+++ docs/changelog.md
@@ -22,7 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
             - Blog page
          - Default "no post found" component
       - Blog sidebar component
-      
+   - Blog controller
 
 ## [1.2.0] - 2021-04-08
 ### Changed